🎨 created calibration class which stores all necessary device information
