+++ tests/ml/test_helper.py
@@ -70,8 +70,6 @@ def test_calc_eval_score_for_qc():
     qc = benchmark_generator.get_benchmark("dj", 1, 3)
     compilation_pipeline = ml.helper.get_compilation_pipeline()
 
-    reward.init_all_config_files()
-
     filename_qasm = "eval_test.qasm"
     for gate_set_name, devices in compilation_pipeline.get("devices").items():
         for device_name, max_qubits in devices:
